icanhelp/requirements.txt
djangorestframework
requests
markdown
django-filter
daphne
channels
asgiref>=3.8.1
django-cors-headers
djangorestframework-simplejwt
gunicorn
whitenoise
python-dotenv
mysqlclient
pillow
django-storages
boto3
django-allauth[socialaccount]
dj-rest-auth[with_social]